REST_Server/api/requirements.txt
asgiref==3.2.3
cffi==1.13.2
cryptography==2.8
defusedxml==0.6.0
Django==3.0.3
djangorestframework==3.11.0
freetype-py==2.1.0.post1
lxml==4.5.0
mysqlclient==1.4.6
numpy==1.17.1
Pillow==7.0.0
pycparser==2.19
PyMySQL==0.9.3
pytz==2019.3
qrcode==6.1
six==1.13.0
sqlparse==0.3.0
